switched custom authentication to identity

+++ WordSnapWeb/WordSnapWeb/Controllers/CardController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using WordSnapWeb.Models;
 
 namespace WordSnapWeb.Controllers
@@ -6,12 +8,15 @@ namespace WordSnapWeb.Controllers
     public class CardController : Controller
     {
         private readonly IWordSnapRepository _repository;
+        private readonly UserManager<ApplicationUser> _users;
 
-        public CardController(IWordSnapRepository repository)
+        public CardController(IWordSnapRepository repository, UserManager<ApplicationUser> users)
         {
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+            _users = users ?? throw new ArgumentNullException(nameof(users));
         }
 
+        [Authorize]
         [HttpGet("EditCard/{cardId}")]
         public async Task<IActionResult> EditCard(int cardId)
         {
@@ -20,10 +25,17 @@ namespace WordSnapWeb.Controllers
             {
                 return NotFound();
             }
-
-            return View(card);
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
+            var cardset = await _repository.GetCardsetByIdAsync(card.CardsetRef);
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                return View(card);
+            }
+            return Unauthorized();
         }
 
+        [Authorize]
         [HttpPost("EditCard/{cardId}")]
         public async Task<IActionResult> EditCard(int cardId, Card updatedCard)
         {
@@ -38,15 +50,23 @@ namespace WordSnapWeb.Controllers
                 return NotFound();
             }
 
-            card.WordEn = updatedCard.WordEn;
-            card.WordUa = updatedCard.WordUa;
-            card.Comment = updatedCard.Comment;
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
+            var cardset = await _repository.GetCardsetByIdAsync(card.CardsetRef);
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                card.WordEn = updatedCard.WordEn;
+                card.WordUa = updatedCard.WordUa;
+                card.Comment = updatedCard.Comment;
 
-            await _repository.UpdateCardAsync(card);
+                await _repository.UpdateCardAsync(card);
 
-            return RedirectToAction("Details", "Cardset", new { cardsetId = card.CardsetRef });
+                return RedirectToAction("Details", "Cardset", new { cardsetId = card.CardsetRef });
+            }
+            return Unauthorized();
         }
 
+        [Authorize]
         [HttpPost("DeleteCard/{cardId}")]
         public async Task<IActionResult> DeleteCard(int cardId)
         {
@@ -56,11 +76,16 @@ namespace WordSnapWeb.Controllers
                 return NotFound();
             }
 
-            int cardsetId = card.CardsetRef;
-
-            await _repository.DeleteCardAsync(cardId);
-
-            return RedirectToAction("Details", "Cardset", new { cardsetId });
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
+            var cardset = await _repository.GetCardsetByIdAsync(card.CardsetRef);
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                int cardsetId = card.CardsetRef;
+                await _repository.DeleteCardAsync(cardId);
+                return RedirectToAction("Details", "Cardset", new { cardsetId });
+            }
+            return Unauthorized();
         }
 
         public IActionResult Index()
