+++ WordSnapWeb/WordSnapWeb/Controllers/CardsetController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Diagnostics;
+using System.Runtime.InteropServices;
 using WordSnapWeb.Models;
-using WordSnapWeb.Services;
 
 namespace WordSnapWeb.Controllers
 {
@@ -9,11 +11,14 @@ namespace WordSnapWeb.Controllers
     public class CardsetController : Controller
     {
         private readonly IWordSnapRepository _repository;
+        private readonly UserManager<ApplicationUser> _users;
+        private readonly SignInManager<ApplicationUser> _signInManager;
 
-
-        public CardsetController(IWordSnapRepository repository)
+        public CardsetController(IWordSnapRepository repository, UserManager<ApplicationUser> users, SignInManager<ApplicationUser> signInManager)
         {
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+            _users = users ?? throw new ArgumentNullException(nameof(users));
+            _signInManager = signInManager ?? throw new ArgumentNullException(nameof(signInManager));
         }
 
         [HttpGet("{cardsetId}")]
@@ -24,9 +29,9 @@ namespace WordSnapWeb.Controllers
             {
                 return NotFound();
             }
-            if (UserSession.Instance.IsLoggedIn)
+            if (_signInManager.IsSignedIn(User))
             {
-                var userscardset = await _repository.GetUserscardsetAsync(UserSession.Instance.CurrentUser.Id, cardsetId);
+                var userscardset = await _repository.GetUserscardsetAsync(_users.GetUserId(User), cardsetId);
                 ViewBag.IsSaved = (userscardset != null);
             }
             else
@@ -37,23 +42,33 @@ namespace WordSnapWeb.Controllers
             return View(cardset);
         }
 
+        [Authorize]
         [HttpPost("{cardsetId}/AddCard")]
         public async Task<IActionResult> CreateCard(int cardsetId, Card card)
         {
             card.CardsetRef = cardsetId;
-            await _repository.AddCardAsync(card);
-            return RedirectToAction("Details", new { cardsetId });
+            var cardset = await _repository.GetCardsetByIdAsync(cardsetId);
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
+
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                await _repository.AddCardAsync(card);
+                return RedirectToAction("Details", new { cardsetId });
+            }
+            return Unauthorized();
         }
 
+        [Authorize]
         [HttpPost("AddToLibrary/{cardsetId}")]
         public async Task<IActionResult> AddToLibrary(int cardsetId)
         {
-            var existing = await _repository.GetUserscardsetAsync(UserSession.Instance.CurrentUser.Id, cardsetId);
+            var existing = await _repository.GetUserscardsetAsync(_users.GetUserId(User), cardsetId);
             if (existing == null)
             {
                 var userscardset = new Userscardset
                 {
-                    UserRef = UserSession.Instance.CurrentUser.Id,
+                    UserRef = _users.GetUserId(User),
                     CardsetRef = cardsetId
                 };
                 await _repository.AddCardsetToSavedLibraryAsync(userscardset);
@@ -61,10 +76,11 @@ namespace WordSnapWeb.Controllers
             return RedirectToAction("Details", new { cardsetId });
         }
 
+        [Authorize]
         [HttpPost("RemoveFromLibrary/{cardsetId}")]
         public async Task<IActionResult> RemoveFromLibrary(int cardsetId)
         {
-            var existing = await _repository.GetUserscardsetAsync(UserSession.Instance.CurrentUser.Id, cardsetId);
+            var existing = await _repository.GetUserscardsetAsync(_users.GetUserId(User), cardsetId);
             if (existing != null)
             {
                 await _repository.DeleteUsersCardset(existing.Id);
@@ -72,6 +88,7 @@ namespace WordSnapWeb.Controllers
             return RedirectToAction("Details", new { cardsetId });
         }
 
+        [Authorize]
         [HttpPost("EditCardset")]
         public async Task<IActionResult> EditCardset(Cardset updatedCardset)
         {
@@ -81,21 +98,38 @@ namespace WordSnapWeb.Controllers
                 return NotFound();
             }
 
-            cardset.Name = updatedCardset.Name;
-            cardset.IsPublic = updatedCardset.IsPublic;
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
 
-            await _repository.UpdateCardsetAsync(cardset);
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                cardset.Name = updatedCardset.Name;
+                cardset.IsPublic = updatedCardset.IsPublic;
 
-            return RedirectToAction("Details", new { cardsetId = cardset.Id });
+                await _repository.UpdateCardsetAsync(cardset);
+
+                return RedirectToAction("Details", new { cardsetId = cardset.Id });
+            }
+            return Unauthorized();
         }
 
+        [Authorize]
         [HttpPost("DeleteCardset")]
         public async Task<IActionResult> DeleteCardset(int cardsetId)
         {
-            await _repository.DeleteCardsetAsync(cardsetId);
-            return RedirectToAction("Index", "Home");
+            var cardset = await _repository.GetCardsetByIdAsync(cardsetId);
+            var isAdmin = User.IsInRole("Admin");
+            var userId = _users.GetUserId(User);
+
+            if (cardset.UserRef == userId || isAdmin)
+            {
+                await _repository.DeleteCardsetAsync(cardsetId);
+                return RedirectToAction("Index", "Home");
+            }
+            return Unauthorized();
         }
 
+        [Authorize]
         [HttpGet("{cardsetId}/TakeTest")]
         public async Task<IActionResult> TakeTest(int cardsetId)
         {
@@ -108,7 +142,7 @@ namespace WordSnapWeb.Controllers
                 TempData["Error"] = "Немає карток для тесту.";
                 return RedirectToAction("Details", new { cardsetId });
             }
-            var progress = await _repository.GetProgress(UserSession.Instance.CurrentUser.Id, cardsetId);
+            var progress = await _repository.GetProgress(_users.GetUserId(User), cardsetId);
             double bestScore = 0;
             if (progress != null)
             {
@@ -125,15 +159,16 @@ namespace WordSnapWeb.Controllers
             return View(model);
         }
 
+        [Authorize]
         [HttpPost("{cardsetId}/TakeTest")]
         public async Task<IActionResult> TakeTest(int cardsetId, double score)
         {
-            var progress = await _repository.GetProgress(UserSession.Instance.CurrentUser.Id, cardsetId);
+            var progress = await _repository.GetProgress(_users.GetUserId(User), cardsetId);
             if (progress == null)
             {
                 var newProgress = new Progress
                 {
-                    UserRef = UserSession.Instance.CurrentUser.Id,
+                    UserRef = _users.GetUserId(User),
                     CardsetRef = cardsetId,
                     SuccessRate = score,
                     LastAccessed = DateTime.Now
